# docs/swagger.yaml
basePath: /api
definitions:
  models.APIError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Invalid request
        type: string
    type: object
  models.Coach:
    properties:
      id:
        type: string
      name:
        type: string
      relationships:
        $ref: '#/definitions/models.CoachRelations'
      userId:
        type: string
    type: object
  models.CoachRelations:
    properties:
      users:
        $ref: '#/definitions/models.Users'
    type: object
  models.Exercises:
    properties:
      category:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Users:
    properties:
      coachId:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      relationships:
        $ref: '#/definitions/models.UsersRelations'
      role:
        type: string
    type: object
  models.UsersRelations:
    properties:
      coach:
        $ref: '#/definitions/models.Coach'
    type: object
  models.Workouts:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      notes:
        type: string
      relationships:
        $ref: '#/definitions/models.WorkoutsRelations'
      scheduledAt:
        type: string
      userId:
        type: string
    type: object
  models.WorkoutsRelations:
    properties:
      Users:
        $ref: '#/definitions/models.Users'
    type: object
info:
  contact:
    name: DoctorVeRossi
    url: https://github.com/vrstelios/GoAPIBackEnd
  description: A Tag service API in Go using Gin framework
  title: EndPoints
  version: "1.3"
paths:
  /Coach/{name}:
    get:
      parameters:
      - description: Coach Name
        in: path
        name: Name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Coach'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get Coach.
      tags:
      - Coach
  /auth/login:
    post:
      description: Authenticates user and returns session cookies
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIError'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Logout user.
      tags:
      - auth
  /auth/register:
    post:
      description: Creates a new user account with username and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/models.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Register a new user.
      tags:
      - auth
  /coach:
    post:
      parameters:
      - description: Coach
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Coach'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIError'
        "208":
          description: Already Reported
          schema:
            $ref: '#/definitions/models.APIError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Coach.
      tags:
      - Coach
  /exercises:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exercises'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Query Exercises.
      tags:
      - Exercise
    post:
      parameters:
      - description: Exercise
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Exercises'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIError'
        "208":
          description: Already Reported
          schema:
            $ref: '#/definitions/models.APIError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Exercise.
      tags:
      - Exercise
  /exercises/{id}:
    get:
      parameters:
      - description: Exercise Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exercises'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get Exercise.
      tags:
      - Exercise
  /workouts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workouts'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Query Workouts.
      tags:
      - Workouts
    post:
      parameters:
      - description: Workout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Workouts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIError'
        "208":
          description: Already Reported
          schema:
            $ref: '#/definitions/models.APIError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Workout.
      tags:
      - Workouts
  /workouts/{id}:
    get:
      parameters:
      - description: Workouts Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workouts'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get Workout.
      tags:
      - Workouts
    put:
      parameters:
      - description: Workout
        in: path
        name: id
        type: string
      - description: Workouts
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Workouts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workouts'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Workout.
      tags:
      - Workouts
swagger: "2.0"